Have FlashHash inherit from Hash
After a long pairing session with Brandon, we figured out why we were getting the marshal load errors in prod.

ArgumentError: dump format error (user class)

The reason was that FlashHash was not actually a hash, but inherits from Hash in Rails 3.  Thus, it blew up in rails 2.

My patch patches up this library to ensure that FlashHash gets defined as a subclass of Hash.